+++ CHANGELOG.rst
@@ -7,6 +7,9 @@
 
 - add ``token-list`` command
 
+- allow ``root`` or users from ``--restrict-modify`` to create tokens for
+  other users
+
 
 0.3.0 (Unreleased)
 ==================